feature/services - customer following migrated to services architecture, added fuzzy search to search mongodb query and amended levenshtein to reflect
