+++ api/cust/deals/search.js
@@ -11,6 +11,7 @@ import { searchFeedSchema } from '#app/validation/customer/deal.js'
 import Err from '#app/services/error/index.js'
 import Task from '#app/services/worker/index.js'
 import DB from '#app/services/db/index.js'
+import { orderSearchDealsByTextMatchRelevance } from '#app/services/worker/tasks/deals.js'
 
 router.get('/', authWithCache, validate(searchFeedSchema), async (req, res) => {
   const {
@@ -24,7 +25,8 @@ router.get('/', authWithCache, validate(searchFeedSchema), async (req, res) => {
 
     const results = await DB.CGetSearchFeed(user, LONG, LAT, text)
 
-    const sorted = await Task.orderSearchDealsByTextMatchRelevance(results, text)
+    // const sorted = await Task.orderSearchDealsByTextMatchRelevance(results, text)
+    const sorted = orderSearchDealsByTextMatchRelevance(JSON.stringify(results), text)
 
     return res.json({ nextCursor: undefined, deals: sorted })
   } catch (error) {
