+++ services/crons/handlers/expire.js
@@ -1,10 +1,8 @@
 import { formatInTimeZone } from 'date-fns-tz'
 
-import Deal from '#app/models/Deal.js'
-import Location from '#app/models/Location.js'
-
 import Err from '#app/services/error/index.js'
 import Mixpanel from '#app/services/mixpanel/index.js'
+import DB from '#app/services/db/index.js'
 
 export const createExpiryDate = () => {
   return formatInTimeZone(new Date(), 'Etc/GMT+12', 'yyyy-MM-dd')
@@ -12,21 +10,9 @@ export const createExpiryDate = () => {
 
 const expireDeals = async () => {
   try {
-    const filter = { end_date: { $lte: createExpiryDate() }, is_expired: false }
-
-    const toExpire = await Deal.find(filter)
-
-    const proms = toExpire.map((deal) =>
-      Location.updateMany(
-        {
-          'restaurant.id': deal.restaurant.id,
-        },
-        { $pull: { active_deals: deal._id } }
-      )
-    )
-    await Promise.all(proms)
-    const expiredReq = await Deal.updateMany(filter, { is_expired: true })
-    Mixpanel.trackExpiredDeals(expiredReq)
+    const expireFrom = createExpiryDate()
+    const expiredResults = await DB.RBulkExpireDealsFromDate(expireFrom)
+    Mixpanel.trackExpiredDeals(expiredResults)
   } catch (error) {
     Err.log(error)
   }
